Modified DelegateSearch Behaviour
Before, when listing delegates for a given course in a given
executionYear, the last active delegates for the period were selected and
shown.
Now, in the current year, delegates with activity periods overlapping
the current time and date are shown, in previous years, delegates with
activity periods overlapping the time and date of the end of the
executionYear are shown.